+++ common-db/src/main/java/cn/featherfly/common/db/builder/dml/basic/SqlJoinOnBuilder.java
@@ -1,0 +1,20 @@
+
+/*
+ * All rights Reserved, Designed By zhongj
+ * @Description:
+ * @author: zhongj
+ * @date: 2024-04-02 17:17:02
+ * @Copyright: 2024 www.featherfly.cn Inc. All rights reserved.
+ */
+package cn.featherfly.common.db.builder.dml.basic;
+
+import cn.featherfly.common.db.builder.SqlBuilder;
+
+/**
+ * SqlJoinOnBuilder.
+ *
+ * @author zhongj
+ */
+public interface SqlJoinOnBuilder extends SqlBuilder {
+
+}